FPGA run still giving problems. PC-Emulrunning when non tester

+++ software/src/iob_versat_csrs_pc_emul.c
@@ -16,4 +16,4 @@ void iob_versat_csrs_init_baseaddr(uint32_t addr) { base = addr; }
 // Core Setters and Getters
 uint32_t iob_versat_csrs_get_interface(int addr) { return 0; }
 
-uint16_t iob_versat_csrs_get_version() { return 0; }
+uint32_t iob_versat_csrs_get_version() { return 0; }
